fix: ruleMatcher circular dependency
