+++ browser-extension/mv3/src/utils.ts
@@ -1,6 +1,4 @@
-import { SourceKey, SourceOperator, UrlSource } from "common/types";
 import config from "common/config";
-import { matchSourceUrl } from "./service-worker/services/ruleMatcher";
 
 export const formatDate = (dateInMillis: number, format: string): string => {
   if (dateInMillis && format === "yyyy-mm-dd") {
@@ -24,23 +22,6 @@ export const getAllSupportedWebURLs = () => {
 
 export const isAppURL = (url: string) => {
   return !!url && getAllSupportedWebURLs().some((webURL) => url.includes(webURL));
-};
-
-export const isBlacklistedURL = (url: string): boolean => {
-  const blacklistedSources: UrlSource[] = [
-    ...getAllSupportedWebURLs().map((webUrl) => ({
-      key: SourceKey.URL,
-      operator: SourceOperator.CONTAINS,
-      value: webUrl,
-    })),
-    {
-      key: SourceKey.URL,
-      operator: SourceOperator.CONTAINS,
-      value: "__rq", // you can use __rq in the url to blacklist it
-    },
-  ];
-
-  return blacklistedSources.some((source) => matchSourceUrl(source, url));
 };
 
 export const generateUrlPattern = (urlString: string) => {
